requirements.txt
azure-ai-agentserver-agentframework==1.0.0b17
agent-framework-core==1.0.0rc3
agent-framework-azure-ai==1.0.0rc3
azure-ai-projects==2.0.0b4
pytest==8.4.2
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.8.1